5. Added a 3rd Screen with 2 screens,
This is the 3. MainPage.dart. Added BottomNavigation that is controlled by widgetList.

The BottomNavigation has 5 widgets, however, currently implented Home and implementing the Analytics page.

The widgetList assists the stateful widget in MainPage. The click of the button updates the currentIndex of the page, which is thus called by the body widget.

This body widget is responsible for the changing of pages, from page to page, with the navigation bar.

The flow of the program currently is:
Hero Section-> Login-> Home -> Analytics/ Other Pages...